requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0 